flatten and n independent trials
adds FlatCmp enum (for Ord types)
adds FlatType enum (for Ord + Add types)

adds flat_cmp and flatten methods to Nested trait.
This allows for controlling which flatten behaviour
should happen with a variable.

allows MapRandVar to run n independent trials for
a given n. This requires one of the two types of
flattening types passed in because otherwise the
return type would be unknown (it would depend on n).

reworks the NumRandVar multiple function to reduce
the number of calls to add_rv.
This is the standard trick for exponentials where
you split n into (n/2) + (n/2) or (n//2) + (n//2) + 1
in a recursive manner.